dags/requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
requests==2.31.0
psycopg2-binary==2.9.9
prefect==2.14.0